Link-ifying Configuration
It really reads like it should be a link next to all the others. Oddly enough there's [not a](http://docs.rancher.com/rancher/latest/en/concepts/#configuration) satisfactory [target link](http://docs.rancher.com/rancher/latest/en/configuration/). The former opens the left nav to present the potential options, but still presents the user at the beginning of the Concepts section. The latter leads to a 404 error. I chose to lead to the Access Control page because it is the first of the options, and as I have just shown there is no good landing spot for the Configuration page at the moment. It seemed like the least evil because it is less likely be mistaken for the Concepts page itself and provides the reader a good starting point to consuming the Configuration section's material - which is not really in a particular order anyway. I understand if this ambiguity makes it weird enough not to link, letting people find it on their own. Ideally it would be corrected by me MR'ing the docs to have the configuration URL not 404 but it did not hit me until I am typing this now. :x