Update index.md
When any service try to use `chown` command in a NFS storage and the NSF server is not mounted with the option `no_root_squash`, it will raised error: 'chown: /var/lib/postgresql/data/pgdata: Operation not permitted'

For example, Postgres service raises this error. With this options this is fixed.